requirements.txt
asgiref==3.7.2
colorama==0.4.6
Django==4.2.8
django-randomfilenamestorage==1.1
djangorestframework==3.14.0
MarkupSafe==2.1.3
pdfkit==1.0.0
pypng==0.20220715.0
python-dotenv==1.0.0
pytz==2023.3.post1
qrcode==7.4.2
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
